# Check Content-Length against minSize and passthrough writes if no gzip