@fastest963 2018-08-08
I also handled the issue mentioned in #64. 

The benchmarks seem to show the `BenchmarkGzipHandler_S2k-8` got faster while the `BenchmarkGzipHandler_P20k-8` and `BenchmarkGzipHandler_P100k-8` got slower.
```
Before:
BenchmarkGzipHandler_S2k-8     	     200	   8180495 ns/op
BenchmarkGzipHandler_S20k-8    	     100	  24256560 ns/op
BenchmarkGzipHandler_S100k-8   	      20	 100509175 ns/op
BenchmarkGzipHandler_P2k-8     	    1000	   1766909 ns/op
BenchmarkGzipHandler_P20k-8    	     200	   5878452 ns/op
BenchmarkGzipHandler_P100k-8   	      50	  26159506 ns/op

After:
BenchmarkGzipHandler_S2k-8     	     200	   6932798 ns/op
BenchmarkGzipHandler_S20k-8    	     100	  24246411 ns/op
BenchmarkGzipHandler_S100k-8   	      20	 101508020 ns/op
BenchmarkGzipHandler_P2k-8     	    1000	   1691251 ns/op
BenchmarkGzipHandler_P20k-8    	     200	   7326576 ns/op
BenchmarkGzipHandler_P100k-8   	      30	  33444100 ns/op
```